Replace all creation of Sequence<byte> except in SequencePool
This promotes much better recycling of arrays and ensures that all arrays have a minimum length so we don't get hundreds of tiny arrays for some users.

I also add a BannedSymbols.txt file so that any attempt to accidentally use "new Sequence<byte>()" will result in a compiler error/warning with advice on what to use instead.
I do this instead of simply removing the API or making any other change to the `Sequence<T>` class so it can stay nearly identical with the implementation found in Nerdbank.Streams, so we can keep it up to date as fixes are made to the original one.

Fixes #688