Revert incremental install of SDKs/Runtimes
The 0c212b6459c4 commit brought this optimization in, but it is buggy because an SDK or runtime might not be installed because it's at the machine level, but if another runtime is missing and installed, it sets env vars which hide the machine level ones already found, so now they need to also be installed. But they weren't, so the build would just fail.

Similarly, if the script is supposed to install to the machine location and the SDK is already installed at the user location, the recent optimization would skip installing it at the machine location.